Changle Template class to Templatable module
